iot_smart-agri/requirements.txt
paho-mqtt==2.1.0
confluent-kafka==2.5.3
pandas==2.2.2
python-dotenv==1.0.1
